Updated docker builds to use the latest stable alpine image
Previously it used the edge distro and the edge/testing apk
repository as a shortcut to installing daemontools. Now we
use the stable version, at the expense of having to build
daemontools from scratch.

We also look up all the tags for alpine:latest, and apply them
to the djbdns image as well. This allows choosy users to pick
what version of alpine they want their image to be based on.

Next up: GitHub actions to build new images whenever the base
image changes.