This is a significant update.
The main addition is a new feature whereby the environment variable
ALLOWED_NETWORKS can be set to a colon-delimited list of (partial)
IP addresses, like this:

ALLOWED_NETWORKS=192.168.1:10.1:172

These are interpreted as if the remaining octets were zeros, and
the netmask was a multiple of 8, so "10.1" means the same thing as
10.1.0.0/16.

Connections from any of those IP ranges will be accepted by the DNS
cache and receive a response. Connections from any other IP address
will receive no response until its connection times out.

In addition, both daemontools and djbdns are now patched, using
patches cribbed from Debian, so that they build successfully on
64-bit systems.